# toEncoding & JSON keys as strings