Hello!

My scenario is JSON serialization with `toEncoding` (to `ByteString`) which results in JSON files
with lots of constructor labels (= lots of keys). I am using `deriveToJSON` to derive the `ToJSON` instances.

The current way of implementing `toEncoding` of objects is by supplying key-value pairs, where the key is an Haskell `String`.
Using strings when serializing to `ByteString`s seems inefficient, so I experimented with using `ByteString` for keys instead.
I only touched the Template Haskell part in such a way that, when deriving `toEncoding`, the labels of constructors are escaped at compile time, and also manually lifted at the toplevel of `toEncoding` and thus shared in subsequent calls to `toEncoding` (see dumped splices below).

I benchmarked an artificial example that builds a big binary tree and serializes it with `toEncoding`:

```haskell
data BinTree = TreeLeaf | TreeNode { leftTree :: BinTree, rightTree :: BinTree }
$(deriveToJSON defaultOptions ''BinTree)
```

The performance improvement is promising (it's more than twice as fast):

- Current `aeson`:
```
time                 1.031 s    (1.026 s .. 1.042 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.027 s    (1.026 s .. 1.029 s)
std dev              2.126 ms   (75.05 μs .. 2.567 ms)
variance introduced by outliers: 19% (moderately inflated)
```

```
70,748,400,904 bytes allocated in the heap
    62,214,640 bytes copied during GC
        158,952 bytes maximum residency (8 sample(s))
        54,320 bytes maximum slop
              0 MB total memory in use (0 MB lost due to fragmentation)
```

- This PR:
```
time                 331.7 ms   (329.8 ms .. 335.1 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 330.0 ms   (329.6 ms .. 331.2 ms)
std dev              903.6 μs   (19.72 μs .. 1.248 ms)
variance introduced by outliers: 16% (moderately inflated)
```

```
44,980,570,088 bytes allocated in the heap
    21,969,912 bytes copied during GC
        157,424 bytes maximum residency (6 sample(s))
        52,760 bytes maximum slop
              0 MB total memory in use (0 MB lost due to fragmentation)
```

(The code for this benchmark is here: https://github.com/acondolu/aeson-bintree)

This is the output of `-ddump-splices`:
```haskell
  deriveToJSON defaultOptions ''BinTree
======>
  instance ToJSON BinTree where
    toJSON = [...]
    toEncoding
      = let
          bs_TreeLeaf = pack [...] -- <-- "\"TreeLeaf\""
          bs_TreeNode = [...]
          bs_leftTree = [...]
          bs_rightTree = [...]
          bs_tag = [...]
        in
          \ value_acGC
            -> case value_acGC of
                  TreeLeaf
                    -> fromPairs
                        ((pair' (Encoding (shortByteString bs_tag)))
                            (Encoding (shortByteString bs_TreeLeaf)))
                  TreeNode arg1_acGD arg2_acGE
                    -> fromPairs
                        ((pair' (Encoding (shortByteString bs_tag)))
                            (Encoding (shortByteString bs_TreeNode))
                            <>
                              ((pair' (Encoding (shortByteString bs_leftTree)))
                                (toEncoding arg1_acGD)
                                <>
                                  (pair' (Encoding (shortByteString bs_rightTree)))
                                    (toEncoding arg2_acGE)))
```

Comparison using `aeson` benchmarks:
```
Benchmark                              master  bs-keys         
Examples/decode/github-issues/lazy    2.91e-3  2.78e-3   -4.73%
Examples/decode/github-issues/strict  3.55e-3  2.71e-3  -23.55%
Examples/decode/jp100/lazy            3.63e-3  3.47e-3   -4.43%
Examples/decode/jp100/strict          3.52e-3  3.51e-3   -0.42%
Examples/decode/twitter100/lazy       2.86e-3  2.77e-3   -3.20%
Examples/decode/twitter100/strict     2.58e-3  2.49e-3   -3.33%
zzz-geom-mean                         3.15e-3  2.93e-3   -6.96%
```

### Alternative solution
This PR only improves the performance when deriving instances with TH (unsatisfactory, for instance `genericToEncoding` is left out).

An alternative would be to leave the TH part unchanged, make `pair` to pack the escaped keys to `ByteString`s, and force GHC to do the lifting by itself, by making it realize that escaped keys are constant expressions after all. However, I was not able to do that:
maybe somebody more experienced with the GHC simplifier can help.

Let me know what you think!