@acondolu 2020-12-17
Hello,

any comment on this PR? Is this performance improvement desirable?

Let me know what I can do to move this PR in some direction.