@phadej 2020-12-18
I also wonder whether it makes sense to let-float the `shortByteString (pack ...)`, not only `pack`. Whether there is difference.

If I were to write something like that by hand, I'd float-out complete `Encoding 'Text` values for keys, i.e.

```
toEncoding =
  let keyA = Encoding.string "a" -- and these could be pre-calced a lot.
      keyB = Encoding.string "b" --
  in \v -> case v of
     ... -> fromPairs (pair' keyA ... <> pair' keyB ...)
```

In fact, the *in theory*, if we construct
https://hackage.haskell.org/package/bytestring-0.11.0.0/docs/Data-ByteString-Builder-Prim.html#t:FixedPrim
(as we know the length!), encoding should go as fast as it could:
one length check for whole key, and then just writing the key data into the
buffer. Nothing for GHC really to optimize there.

---

Could you experiment whether that is viable, and have an effect.
Reducing how much stuff GHC optimizer have to consider has effect
on compilation time (especially if people go crazy and use `{-# INLINE #-}`
everywhere).

If we have `FixedPrim` stuff, it would be up to GHC to decide whether they are cheap to inline, or keep floated-out.
And we should let GHC do that. Inlining is cheaper optimization to perform than CSE.

---

I can then think how to get rid of `IORef` in elegant way.
(I know how, but it's indeed not a one-line patch, though quite
straigh-forward anyway).