@acondolu 2020-12-21
Hi,
I've applied the change suggested above (floating the `Encoding`s instead of the `ByteString`s). I've carried again the benchmark and the performance gain remains, so GHC is still happy. 🎉

About the `FixedPrim` suggestion: I think what `shortByteString :: ShortByteString -> Builder` does is basically what mentioned by @phadej in the comment above? i.e. checking the limits of the buffer and then copying the underlying byte array (see https://hackage.haskell.org/package/bytestring-0.11.0.0/docs/src/Data.ByteString.Builder.Internal.html#shortByteString). (The important thing is ensuring that the `ShortByteString` is computed only once, and not at every call of `toEncoding`. Once it's computed, I agree that there's nothing to optimise and it's as fast as possible).