@phadej 2021-12-07
I implemented this idea in https://github.com/haskell/aeson/pull/906